+++ noderedpy/__init__.py
@@ -1,17 +1,21 @@
 # -*- coding: utf-8 -*-
-from ._nodered import RED, NodeCommunicator as Node
-from ._property import (
+"""
+make python function to Node-RED node
+"""
+
+from .__nodered__ import RED, REDBuilder, NodeCommunicator as Node
+from .__nodered__.auth import Auth
+from .__property__ import (
     InputProperty, ListProperty, DictProperty, CodeProperty,
     SpinnerProperty, CheckBoxProperty, ComboBoxProperty
 )
 
-
-__version__ = "0.2.8"
+__version__ = "0.2.9"
 __author__ = "oyajiDev"
 __email__ = "this.dev.somehit@gmail.com"
 
 __all__ = [
-    "RED", "Node",
+    "RED", "REDBuilder", "Auth", "Node",
     "InputProperty", "ListProperty", "DictProperty", "CodeProperty",
     "SpinnerProperty", "CheckBoxProperty", "ComboBoxProperty"
 ]